refactor(netcore): separate resolver and dialer code
No functional change, just more clear separation of concerns.
