+++ netcore/tlsconfig.go
@@ -1,4 +1,8 @@
+//
 // SPDX-License-Identifier: GPL-3.0-or-later
+//
+// TLS config code.
+//
 
 package netcore
