+++ netcore/tlsdialer.go
@@ -1,4 +1,8 @@
+//
 // SPDX-License-Identifier: GPL-3.0-or-later
+//
+// TLS dialing code
+//
 
 package netcore
